# openapi.yaml
openapi: 3.1.0
info:
  title: PM Assistant API
  version: 1.0.0
  description: >
    REST API สำหรับผู้ช่วย PM บน LINE — ใช้โดยหน้าเว็บตั้งค่า (LIFF) และรับ event จาก
    LINE/แพลตฟอร์มภายนอก. ใช้คู่กับ schema.sql และ DESIGN_SPEC.md.
    ทุก endpoint ฝั่ง LIFF ต้องมี bearer token (จาก LINE Login); endpoint /webhooks
    ตรวจ signature แทน.
servers:
  - url: https://api.pmassistant.app/v1
    description: production
  - url: http://localhost:8000/v1
    description: local dev
...
security:
  - bearerAuth: []
paths:
...
  /projects:
    get:
      tags: [Projects]
      summary: รายการโปรเจกต์ของผู้ใช้
      responses:
        '200':
          description: สำเร็จ
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Project' }
    post:
      tags: [Projects]
      summary: สร้างโปรเจกต์
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProjectCreate' }
      responses:
        '201':
          description: สร้างแล้ว
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Project' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /projects/{projectId}:
    parameters:
      - $ref: '#/components/parameters/projectId'
    get:
      tags: [Projects]
      summary: รายละเอียดโปรเจกต์
      responses:
        '200':
          description: สำเร็จ
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Project' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [Projects]
      summary: แก้ไขโปรเจกต์
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProjectCreate' }
      responses:
        '200':
          description: สำเร็จ
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Project' }
    delete:
      tags: [Projects]
      summary: เก็บถาวร (archive) โปรเจกต์
      responses:
        '204': { description: สำเร็จ }
  /projects/{projectId}/dashboard:
    parameters:
      - $ref: '#/components/parameters/projectId'
    get:
      tags: [Projects]
      summary: ข้อมูลแดชบอร์ดภาพรวม (หน้า W2)
      responses:
        '200':
          description: สำเร็จ
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DashboardSummary' }

  # ---------------------------------------------------------------- Members
...
  /projects/{projectId}/templates:
    parameters:
      - $ref: '#/components/parameters/projectId'
    get:
      tags: [Templates]
      summary: เทมเพลตของโครงการ
      responses:
        '200':
          description: สำเร็จ
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Template' }
    post:
      tags: [Templates]
      summary: อัปโหลดเทมเพลต (.docx)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [wpType, file]
              properties:
                wpType: { $ref: '#/components/schemas/WorkProductType' }
                file: { type: string, format: binary }
      responses:
        '201':
          description: สำเร็จ
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Template' }
  /templates/{templateId}:
    parameters:
      - name: templateId
        in: path
        required: true
        schema: { type: string, format: uuid }
    delete:
      tags: [Templates]
      summary: ลบเทมเพลต
      responses:
        '204': { description: สำเร็จ }

  # -------------------------------------------------------------- Numbering
  /projects/{projectId}/numbering:
    parameters:
      - $ref: '#/components/parameters/projectId'
    get:
      tags: [Numbering]
      summary: กฎเลขรันเอกสารต่อชนิด
      responses:
        '200':
          description: สำเร็จ
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/NumberingRule' }
  /projects/{projectId}/numbering/{wpType}:
    parameters:
      - $ref: '#/components/parameters/projectId'
      - name: wpType
        in: path
        required: true
        schema: { $ref: '#/components/schemas/WorkProductType' }
    put:
      tags: [Numbering]
      summary: ตั้งรูปแบบเลขรัน
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                prefix: { type: string, example: MIN }
                pattern: { type: string, example: '{KEY}-{TYPE}-{SEQ:04d}' }
                resetPeriod: { type: string, enum: [none, yearly] }
      responses:
        '200':
          description: สำเร็จ
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NumberingRule' }

  # -------------------------------------------------------------- Documents
...
  /projects/{projectId}/tasks:
    parameters:
      - $ref: '#/components/parameters/projectId'
      - name: status
        in: query
        schema: { $ref: '#/components/schemas/TaskStatus' }
      - name: dueBefore
        in: query
        schema: { type: string, format: date }
      - name: milestoneId
        in: query
        schema: { type: string, format: uuid }
    get:
      tags: [Tasks]
      summary: รายการงาน (ใช้โดยแดชบอร์ด/ถาม-ตอบ)
      responses:
        '200':
          description: สำเร็จ
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Task' }
    post:
      tags: [Tasks]
      summary: สร้างงาน (ปกติผ่านแชท แต่เปิด REST ไว้ด้วย)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TaskInput' }
      responses:
        '201':
          description: สำเร็จ
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Task' }
  /tasks/{taskId}:
    parameters:
      - name: taskId
        in: path
        required: true
        schema: { type: string, format: uuid }
    patch:
      tags: [Tasks]
      summary: แก้ไขงาน
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TaskInput' }
      responses:
        '200':
          description: สำเร็จ
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Task' }

  # ------------------------------------------------------------- Milestones
  /projects/{projectId}/milestones:
    parameters:
      - $ref: '#/components/parameters/projectId'
    get:
      tags: [Milestones]
      summary: รายการ milestone
      responses:
        '200':
          description: สำเร็จ
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Milestone' }
    post:
      tags: [Milestones]
      summary: สร้าง milestone
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MilestoneInput' }
      responses:
        '201':
          description: สำเร็จ
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Milestone' }

  # --------------------------------------------------------------- Meetings
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: token ที่ออกหลังตรวจ LINE Login ID token จากฝั่ง LIFF
  parameters:
    projectId:
      name: projectId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    NotFound:
      description: ไม่พบข้อมูล
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    ValidationError:
      description: ข้อมูลไม่ผ่านการตรวจสอบ
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
...
    WorkProductType:
      type: string
      enum: [meeting_record, memo, project_plan, requirements, traceability, test_case, change_request]
    TaskStatus:
      type: string
      enum: [todo, in_progress, blocked, done, cancelled]
    TaskPriority:
      type: string
      enum: [low, medium, high, urgent]
    MilestoneStatus:
      type: string
      enum: [open, at_risk, done]
...
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string, example: not_found }
        message: { type: string }
        details: { type: object, additionalProperties: true }

    # ---- projects ----
    Project:
      type: object
      properties:
        id: { type: string, format: uuid }
        key: { type: string, example: HR }
        name: { type: string, example: HR system }
        isArchived: { type: boolean }
        createdAt: { type: string, format: date-time }
    ProjectCreate:
      type: object
      required: [key, name]
      properties:
        key: { type: string, maxLength: 10 }
        name: { type: string }
    DashboardSummary:
      type: object
      properties:
        taskCounts:
          type: object
          properties:
            total: { type: integer }
            pending: { type: integer }
            done: { type: integer }
        nextMilestone: { $ref: '#/components/schemas/Milestone' }
        dueSoon:
          type: array
          items: { $ref: '#/components/schemas/Task' }
        nextEvent: { $ref: '#/components/schemas/CalendarEvent' }

    # ---- members ----
...
    Template:
      type: object
      properties:
        id: { type: string, format: uuid }
        wpType: { $ref: '#/components/schemas/WorkProductType' }
        version: { type: integer }
        uploadedAt: { type: string, format: date-time }
    NumberingRule:
      type: object
      properties:
        wpType: { $ref: '#/components/schemas/WorkProductType' }
        prefix: { type: string }
        pattern: { type: string }
        currentSeq: { type: integer }
        resetPeriod: { type: string, enum: [none, yearly] }

    # ---- documents ----
...
    Task:
      type: object
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        assigneeId: { type: string, format: uuid, nullable: true }
        dueDate: { type: string, format: date, nullable: true }
        priority: { $ref: '#/components/schemas/TaskPriority' }
        status: { $ref: '#/components/schemas/TaskStatus' }
        milestoneId: { type: string, format: uuid, nullable: true }
        externalRef:
          type: object
          description: ลิงก์ไป card บนแพลตฟอร์ม
          properties:
            provider: { type: string }
            key: { type: string, example: HR-128 }
            url: { type: string, format: uri }
    TaskInput:
      type: object
      required: [title]
      properties:
        title: { type: string }
        description: { type: string }
        assigneeId: { type: string, format: uuid }
        dueDate: { type: string, format: date }
        priority: { $ref: '#/components/schemas/TaskPriority' }
        milestoneId: { type: string, format: uuid }
    Milestone:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        targetDate: { type: string, format: date, nullable: true }
        status: { $ref: '#/components/schemas/MilestoneStatus' }
        linkedTaskCount: { type: integer }
    MilestoneInput:
      type: object
      required: [name]
      properties:
        name: { type: string }
        description: { type: string }
        targetDate: { type: string, format: date }

    # ---- meetings ----
...
    CalendarEvent:
      type: object
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        startsAt: { type: string, format: date-time }
        endsAt: { type: string, format: date-time, nullable: true }
        recurrenceRule: { type: string, nullable: true, example: 'FREQ=WEEKLY;BYDAY=MO' }
        meetLink: { type: string, format: uri, nullable: true }
...